+++ src/GitHub_Updater/Bitbucket_API.php
@@ -37,8 +37,7 @@ class Bitbucket_API extends API {
 		parent::$hours  = 12;
 		$this->response = $this->get_transient();
 
-		add_filter( 'http_request_args', array( &$this, 'maybe_authenticate_http' ), 10, 2 );
-		add_filter( 'http_request_args', array( &$this, 'http_release_asset_auth' ), 15, 2 );
+		$this->load_hooks();
 
 		if ( ! isset( self::$options['bitbucket_username'] ) ) {
 			self::$options['bitbucket_username'] = null;
@@ -47,6 +46,21 @@ class Bitbucket_API extends API {
 			self::$options['bitbucket_password'] = null;
 		}
 		add_site_option( 'github_updater', self::$options );
+	}
+
+	public function load_hooks() {
+		add_filter( 'http_request_args', array( &$this, 'maybe_authenticate_http' ), 10, 2 );
+		add_filter( 'http_request_args', array( &$this, 'http_release_asset_auth' ), 15, 2 );
+		add_filter( 'http_request_args', array( &$this, 'ajax_maybe_authenticate_http' ), 15, 2 );
+	}
+
+	/**
+	 * Remove hooks for Bitbucket authentication headers.
+	 */
+	public function remove_hooks() {
+		remove_filter( 'http_request_args', array( &$this, 'maybe_authenticate_http' ) );
+		remove_filter( 'http_request_args', array( &$this, 'http_release_asset_auth' ) );
+		remove_filter( 'http_request_args', array( &$this, 'ajax_maybe_authenticate_http' ) );
 	}
 
 	/**
@@ -90,7 +104,7 @@ class Bitbucket_API extends API {
 		$repo_type = $this->return_repo_type();
 		$response  = isset( $this->response['tags'] ) ? $this->response['tags'] : false;
 
-		if ( $this->exit_no_update( $response ) && 'theme' !== $repo_type['type'] ) {
+		if ( $this->exit_no_update( $response, true ) ) {
 			return false;
 		}
 
@@ -501,7 +515,7 @@ class Bitbucket_API extends API {
 	 *
 	 * @return mixed
 	 */
-	public static function ajax_maybe_authenticate_http( $args, $url ) {
+	public function ajax_maybe_authenticate_http( $args, $url ) {
 		if ( parent::is_doing_ajax() && ! parent::is_heartbeat() &&
 		     ( isset( $_POST['slug'] ) && array_key_exists( $_POST['slug'], parent::$options ) &&
 		       1 == parent::$options[ $_POST['slug'] ] &&