+++ src/GitHub_Updater/GitLab_API.php
@@ -44,14 +44,14 @@ class GitLab_API extends API {
 		parent::$hours  = 12;
 		$this->response = $this->get_transient();
 
-		if ( ! isset( self::$options['gitlab_private_token'] ) ) {
-			self::$options['gitlab_private_token'] = null;
+		if ( ! isset( self::$options['gitlab_access_token'] ) ) {
+			self::$options['gitlab_access_token'] = null;
 		}
 		if ( ! isset( self::$options['gitlab_enterprise_token'] ) ) {
 			self::$options['gitlab_enterprise_token'] = null;
 		}
 		if (
-			empty( self::$options['gitlab_private_token'] ) ||
+			empty( self::$options['gitlab_access_token'] ) ||
 			( empty( self::$options['gitlab_enterprise_token'] ) && ! empty( $type->enterprise ) )
 		) {
 			Messages::instance()->create_error_message( 'gitlab' );
@@ -108,7 +108,7 @@ class GitLab_API extends API {
 		$repo_type = $this->return_repo_type();
 		$response  = isset( $this->response['tags'] ) ? $this->response['tags'] : false;
 
-		if ( $this->exit_no_update( $response ) && 'theme' !== $repo_type['type'] ) {
+		if ( $this->exit_no_update( $response, true ) ) {
 			return false;
 		}
 
@@ -244,6 +244,10 @@ class GitLab_API extends API {
 	public function get_repo_meta() {
 		$response = isset( $this->response['meta'] ) ? $this->response['meta'] : false;
 
+		if ( $this->exit_no_update( $response ) ) {
+			return false;
+		}
+
 		if ( ! $response ) {
 			self::$method = 'meta';
 			$projects     = isset( $this->response['projects'] ) ? $this->response['projects'] : false;
@@ -349,25 +353,26 @@ class GitLab_API extends API {
 		}
 
 		/*
+		 * If a branch has been given, only check that for the remote info.
+		 * If branch is master (default) and tags are used, use newest tag.
+		 */
+		if ( 'master' === $this->type->branch && ! empty( $this->type->tags ) ) {
+			$endpoint = remove_query_arg( 'ref', $endpoint );
+			$endpoint = add_query_arg( 'ref', $this->type->newest_tag, $endpoint );
+		} elseif ( ! empty( $this->type->branch ) ) {
+			$endpoint = remove_query_arg( 'ref', $endpoint );
+			$endpoint = add_query_arg( 'ref', $this->type->branch, $endpoint );
+		}
+
+		/*
 		 * Check for rollback.
 		 */
 		if ( ! empty( $_GET['rollback'] ) &&
 		     ( isset( $_GET['action'] ) && 'upgrade-theme' === $_GET['action'] ) &&
 		     ( isset( $_GET['theme'] ) && $this->type->repo === $_GET['theme'] )
 		) {
-			$endpoint = add_query_arg( 'ref', esc_attr( $_GET['rollback'] ), $endpoint );
-		} elseif ( ! empty( $this->type->branch ) ) {
-			$endpoint = add_query_arg( 'ref', $this->type->branch, $endpoint );
-		}
-
-		/*
-		 * If a branch has been given, only check that for the remote info.
-		 * If it's not been given, GitLab will use the Default branch.
-		 * If branch is master and tags are used, use newest tag.
-		 */
-		if ( 'master' === $this->type->branch && ! empty( $this->type->tags ) ) {
 			$endpoint = remove_query_arg( 'ref', $endpoint );
-			$endpoint = add_query_arg( 'ref', $this->type->newest_tag, $endpoint );
+			$endpoint = add_query_arg( 'ref', esc_attr( $_GET['rollback'] ), $endpoint );
 		}
 
 		/*
@@ -437,18 +442,23 @@ class GitLab_API extends API {
 	 */
 	private function add_access_token_endpoint( $git, $endpoint ) {
 
-		if ( ! empty( parent::$options['gitlab_private_token'] ) ) {
-			$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_private_token'], $endpoint );
+		// Add GitLab.com Access Token.
+		if ( ! empty( parent::$options['gitlab_access_token'] ) ) {
+			$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_access_token'], $endpoint );
 		}
 
-		/*
-		 * If using GitLab CE/Enterprise header return this endpoint.
-		 */
-		if ( ! empty( $git->type->enterprise ) ) {
+		// If using GitLab CE/Enterprise header return this endpoint.
+		if ( ! empty( $git->type->enterprise ) &&
+		     ! empty( parent::$options['gitlab_enterprise_token'] )
+		) {
 			$endpoint = remove_query_arg( 'private_token', $endpoint );
-			if ( ! empty( parent::$options['gitlab_enterprise_token'] ) ) {
-				$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_enterprise_token'], $endpoint );
-			}
+			$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_enterprise_token'], $endpoint );
+		}
+
+		// Add repo access token.
+		if ( ! empty( parent::$options[ $git->type->repo ] ) ) {
+			$endpoint = remove_query_arg( 'private_token', $endpoint );
+			$endpoint = add_query_arg( 'private_token', parent::$options[ $git->type->repo ], $endpoint );
 		}
 
 		return $endpoint;
@@ -474,9 +484,6 @@ class GitLab_API extends API {
 	 * @return string
 	 */
 	protected function add_endpoints( $git, $endpoint ) {
-		if ( ! empty( parent::$options['gitlab_private_token'] ) ) {
-			$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_private_token'], $endpoint );
-		}
 
 		switch ( self::$method ) {
 			case 'projects':
@@ -495,15 +502,12 @@ class GitLab_API extends API {
 				break;
 		}
 
+		$endpoint = $this->add_access_token_endpoint( $git, $endpoint );
+
 		/*
 		 * If using GitLab CE/Enterprise header return this endpoint.
 		 */
 		if ( ! empty( $git->type->enterprise_api ) ) {
-			$endpoint = remove_query_arg( 'private_token', $endpoint );
-			if ( ! empty( parent::$options['gitlab_enterprise_token'] ) ) {
-				$endpoint = add_query_arg( 'private_token', parent::$options['gitlab_enterprise_token'], $endpoint );
-			}
-
 			return $git->type->enterprise_api . $endpoint;
 		}
 
@@ -521,7 +525,7 @@ class GitLab_API extends API {
 
 		if ( ! $response ) {
 			self::$method = 'projects';
-			$response     = $this->api( '/projects' );
+			$response     = $this->api( '/projects?per_page=100' );
 			if ( empty( $response ) ) {
 				$id = urlencode( $this->type->owner . '/' . $this->type->repo );