+++ src/GitHub_Updater/Language_Pack.php
@@ -112,8 +112,5 @@ class Language_Pack extends Base {
 
 		return $transient;
 	}
-<<<<<<< HEAD
-=======
 
->>>>>>> develop
 }